:recycle: move renderer to child component, add search icon styles
