+++ src/components/autoSuggest/index.test.js
@@ -58,11 +58,11 @@ describe('AutoSuggest', () => {
     expect(wrapper.find('.transactions-header')).to.have.lengthOf(0);
   });
 
-  it('should show autosuggest on focus and hide it on blur', () => {
+  it('should show autosuggest on search input change and hide it on blur', () => {
     let autosuggestDropdown = wrapper.find('.autosuggest-dropdown').first();
     const autosuggestInput = wrapper.find('.autosuggest-input').find('input').first();
     expect(autosuggestDropdown.props().className.match(new RegExp(/ {1}autoSuggest__show__/g))).to.be.equal(null);
-    autosuggestInput.simulate('focus');
+    autosuggestInput.simulate('change', { target: { value: 'abc' } });
     wrapper.update();
     autosuggestDropdown = wrapper.find('.autosuggest-dropdown').first();
     expect(autosuggestDropdown.props().className.match(new RegExp(/ {1}autoSuggest__show__/g))).not.to.be.equal(null);
