+++ src/components/autoSuggest/resultList.css
@@ -1,0 +1,50 @@
+@import './../app/variables.css';
+
+.resultList {
+  list-style-type: none;
+  display: flex;
+  flex-direction: column;
+  align-content: flex-start;
+  margin: 0;
+  padding: 0;
+}
+
+.row {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  margin: 0;
+  padding: 0;
+  color: var(--color-primary-standard);
+  font-family: var(--content-font);
+  font-size: var(--paragraph-font-size-l);
+
+  &:hover {
+    background-color: rgba(205, 218, 229, 0.2);
+  }
+
+  & > span {
+    margin: 0 24px;
+    line-height: 36px;
+  }
+}
+
+.rowResult {
+  cursor: pointer;
+}
+
+.rowSelected {
+  background-color: rgba(204, 217, 228, 0.2);
+}
+
+.heading {
+  margin-top: 24px;
+  color: var(--color-grayscale-medium);
+  font-family: var(--content-font);
+  font-size: var(--font-size-small);
+  font-weight: var(--font-weight-bold);
+
+  &:hover {
+    background-color: inherit;
+  }
+}
